Created a create database file

+++ CreateBloodTypeTable.sql
@@ -1,5 +1,5 @@
 CREATE TABLE BloodDonationDB.dbo.BloodType(
-[BloodTypeID] [int] IDENTITY(1,1) NOT NULL,
+[BloodTypeID] [SMALLINT] IDENTITY(1,1) NOT NULL,
 [BloodTypeName] [varchar] NOT NULL,
 [Description] [varchar] NULL,
 PRIMARY KEY ([BloodTypeID])
